Update details about GrimoireLab and added resources
This commit updates the development page with the grimoirelab
details and the list of projects. It also adds the tech stack
for the projects and a few resources for learning them.

Added resources & information about writing a good commit message
and link for DCO sign-off browser plugin.

It also formats the content to align to a particular style.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@chaoss.community>